Move models from _menu_basic to models class for clearer oversight & sharing types between classes
