+++ sous-chef/sous_chef/menu/create_menu/_for_grocery_list.py
@@ -12,10 +12,8 @@ from sous_chef.formatter.ingredient.format_ingredient import (
     IngredientFormatter,
     MapIngredientErrorToException,
 )
-from sous_chef.menu.create_menu._menu_basic import (
-    MenuIncompleteError,
-    TmpMenuSchema,
-)
+from sous_chef.menu.create_menu._menu_basic import MenuIncompleteError
+from sous_chef.menu.create_menu.models import TmpMenuSchema
 from sous_chef.recipe_book.read_recipe_book import RecipeBook
 from sous_chef.recipe_book.recipe_util import (
     MapRecipeErrorToException,
