+++ sous-chef/tests/unit_tests/conftest.py
@@ -7,7 +7,7 @@ import pandas as pd
 import pytest
 from hydra import compose, initialize
 from pandera.typing.common import DataFrameBase
-from sous_chef.menu.create_menu._menu_basic import BasicMenuSchema
+from sous_chef.menu.create_menu.models import BasicMenuSchema
 from sous_chef.menu.record_menu_history import MenuHistorian
 from tests.conftest import FROZEN_DATETIME
 from tests.data.util_data import get_all_menus
