+++ sous-chef/tests/unit_tests/menu/create_menu/test__from_fixed_template.py
@@ -6,7 +6,8 @@ from freezegun import freeze_time
 from sous_chef.formatter.ingredient.format_ingredient import Ingredient
 from sous_chef.formatter.units import unit_registry
 from sous_chef.menu.create_menu._from_fixed_template import FixedTemplates
-from sous_chef.menu.create_menu._menu_basic import MenuFutureError, Season
+from sous_chef.menu.create_menu._menu_basic import MenuFutureError
+from sous_chef.menu.create_menu.models import Season
 from sous_chef.menu.record_menu_history import MenuHistoryError
 from tests.conftest import FROZEN_DATE
 from tests.unit_tests.util import create_recipe
